# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a PDD-GitHub-Go-Backend API server",
        "title": "PDD-GitHub-Go-Backend API",
        "contact": {
            "name": "murasame29",
            "email": "oogiriminister@gamil.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/encounters": {
            "get": {
                "description": "Get All Encounters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Encounter"
                ],
                "summary": "Get All Encounters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "pageID",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "userID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/output.ListEncounterResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "error response"
                    },
                    "500": {
                        "description": "error response"
                    }
                }
            },
            "post": {
                "description": "Create Encount entory",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Encounter"
                ],
                "summary": "Create Encounter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "create encounter request",
                        "name": "CreateEncounterRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/input.CreateEncounterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "$ref": "#/definitions/output.CreateEncounterResponse"
                        }
                    },
                    "400": {
                        "description": "error response"
                    },
                    "500": {
                        "description": "error response"
                    }
                }
            }
        },
        "/encounters/{encounter_id}": {
            "get": {
                "description": "Get All Encounters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Encounter"
                ],
                "summary": "Get All Encounters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "list encounter request",
                        "name": "encounter_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "$ref": "#/definitions/output.ListEncounterResponse"
                        }
                    },
                    "400": {
                        "description": "error response"
                    },
                    "500": {
                        "description": "error response"
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Create User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "create user request",
                        "name": "CreateUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/input.CreateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "$ref": "#/definitions/output.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "error response"
                    },
                    "409": {
                        "description": "error response"
                    },
                    "500": {
                        "description": "error response"
                    }
                }
            }
        },
        "/users/{user_id}": {
            "get": {
                "description": "Get any User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "create user request",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "$ref": "#/definitions/output.ReadUserResponse"
                        }
                    },
                    "400": {
                        "description": "error response"
                    },
                    "409": {
                        "description": "error response"
                    },
                    "500": {
                        "description": "error response"
                    }
                }
            },
            "delete": {
                "description": "Delete User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "create user request",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "$ref": "#/definitions/output.DeleteUserResponse"
                        }
                    },
                    "400": {
                        "description": "error response"
                    },
                    "409": {
                        "description": "error response"
                    },
                    "500": {
                        "description": "error response"
                    }
                }
            }
        }
    },
    "definitions": {
        "input.CreateEncounterRequest": {
            "type": "object",
            "properties": {
                "encounted_user_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "input.CreateUser": {
            "type": "object",
            "properties": {
                "github_id": {
                    "type": "string"
                }
            }
        },
        "output.CreateEncounterResponse": {
            "type": "object",
            "properties": {
                "encounter_id": {
                    "type": "string"
                }
            }
        },
        "output.CreateUserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "ユーザの情報",
                    "type": "string"
                }
            }
        },
        "output.DeleteUserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "output.ListEncounterResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "encounter_id": {
                    "type": "string"
                },
                "encouted_user_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "output.ReadUserResponse": {
            "type": "object",
            "properties": {
                "github_id": {
                    "type": "string"
                },
                "user_id": {
                    "description": "ユーザの情報",
                    "type": "string"
                }
            }
        }
    }
}